# artifacts/api-server/src/routes/storage.ts
import { Router, type IRouter, type Request, type Response } from "express";
import { Readable } from "stream";
import { createReadStream } from "node:fs";
import { stat } from "node:fs/promises";
import path from "node:path";
import {
  RequestUploadUrlBody,
  RequestUploadUrlResponse,
} from "@workspace/api-zod";
import { ObjectStorageService, ObjectNotFoundError } from "../lib/objectStorage";
import { ObjectPermission } from "../lib/objectAcl";
import { db, libraryEntriesTable, shareLinksTable } from "@workspace/db";
import { or, eq } from "drizzle-orm";

// ----------------------- upload authorization -----------------------
//
// /storage/uploads/request-url is the entry point for putting bytes into
// object storage.  We restrict it to either:
//   1. The library owner (LIBRARY_OWNER_TOKEN, sent as Authorization: Bearer
//      or x-library-owner-token).
//   2. A non-revoked, non-expired share link (its random token sent as
//      x-share-token).  Share contributors need this to upload through
//      tokenized links.
//
// Without this guard, anyone could mint signed upload URLs and write
// arbitrary objects to our bucket.

const OWNER_TOKEN = process.env.LIBRARY_OWNER_TOKEN;

function timingSafeEq(a: string, b: string): boolean {
  if (a.length !== b.length) return false;
  const ab = Buffer.from(a, "utf8");
  const bb = Buffer.from(b, "utf8");
  let mismatch = 0;
  for (let i = 0; i < ab.length; i++) mismatch |= ab[i]! ^ bb[i]!;
  return mismatch === 0;
}

function extractOwnerToken(req: Request): string | null {
  const header = req.header("x-library-owner-token");
  if (header) return header.trim();
  const auth = req.header("authorization");
  if (auth && auth.toLowerCase().startsWith("bearer ")) {
    return auth.slice(7).trim();
  }
  return null;
}

async function shareTokenIsValid(token: string): Promise<boolean> {
  if (!token) return false;
  const rows = await db
    .select()
    .from(shareLinksTable)
    .where(eq(shareLinksTable.token, token))
    .limit(1);
  if (!rows.length) return false;
  const link = rows[0]!;
  if (link.revokedAt) return false;
  if (link.expiresAt && link.expiresAt.getTime() <= Date.now()) return false;
  return true;
}

async function requireUploadAuth(
  req: Request,
  res: Response,
  next: () => void,
): Promise<void> {
  const owner = extractOwnerToken(req);
  if (OWNER_TOKEN && owner && timingSafeEq(owner, OWNER_TOKEN)) {
    next();
    return;
  }
  const shareToken = req.header("x-share-token")?.trim();
  if (shareToken && (await shareTokenIsValid(shareToken))) {
    next();
    return;
  }
  res.status(401).json({ error: "Unauthorized" });
}

const ATTACHED_ASSETS_DIR = process.env.REPL_HOME
  ? path.resolve(process.env.REPL_HOME, "attached_assets")
  : path.resolve(process.cwd(), "../../attached_assets");

function guessContentType(name: string): string {
  const ext = name.toLowerCase().split(".").pop() || "";
  switch (ext) {
    case "pdf": return "application/pdf";
    case "png": return "image/png";
    case "jpg":
    case "jpeg": return "image/jpeg";
    case "gif": return "image/gif";
    case "webp": return "image/webp";
    case "svg": return "image/svg+xml";
    case "xlsx": return "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
    case "xls": return "application/vnd.ms-excel";
    case "docx": return "application/vnd.openxmlformats-officedocument.wordprocessingml.document";
    case "doc": return "application/msword";
    case "csv": return "text/csv";
    case "txt": return "text/plain; charset=utf-8";
    case "zip": return "application/zip";
    default: return "application/octet-stream";
  }
}

const router: IRouter = Router();
const objectStorageService = new ObjectStorageService();

/**
 * POST /storage/uploads/request-url
 *
 * Request a presigned URL for file upload.
 * The client sends JSON metadata (name, size, contentType) — NOT the file.
 * Then uploads the file directly to the returned presigned URL.
 */
router.post("/storage/uploads/request-url", requireUploadAuth, async (req: Request, res: Response) => {
  const parsed = RequestUploadUrlBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Missing or invalid required fields" });
    return;
  }

  try {
    const { name, size, contentType } = parsed.data;

    const uploadURL = await objectStorageService.getObjectEntityUploadURL();
    const objectPath = objectStorageService.normalizeObjectEntityPath(uploadURL);

    res.json(
      RequestUploadUrlResponse.parse({
        uploadURL,
        objectPath,
        metadata: { name, size, contentType },
      }),
    );
  } catch (error) {
    req.log.error({ err: error }, "Error generating upload URL");
    res.status(500).json({ error: "Failed to generate upload URL" });
  }
});

/**
 * GET /storage/public-objects/*
 *
 * Serve public assets from PUBLIC_OBJECT_SEARCH_PATHS.
 * These are unconditionally public — no authentication or ACL checks.
 * IMPORTANT: Always provide this endpoint when object storage is set up.
 */
router.get("/storage/public-objects/*filePath", async (req: Request, res: Response) => {
  try {
    const raw = req.params.filePath;
    const filePath = Array.isArray(raw) ? raw.join("/") : raw;

    // Local-filesystem fallback for `attached_assets/...` — these files live in the
    // project's attached_assets/ directory rather than object storage.
    if (filePath.startsWith("attached_assets/")) {
      const rel = filePath.slice("attached_assets/".length);
      // Prevent directory traversal.
      const abs = path.resolve(ATTACHED_ASSETS_DIR, rel);
      if (!abs.startsWith(ATTACHED_ASSETS_DIR + path.sep) && abs !== ATTACHED_ASSETS_DIR) {
        res.status(400).json({ error: "Invalid path" });
        return;
      }
      try {
        const s = await stat(abs);
        if (!s.isFile()) {
          res.status(404).json({ error: "File not found" });
          return;
        }
        res.setHeader("Content-Type", guessContentType(abs));
        res.setHeader("Content-Length", String(s.size));
        res.setHeader("Cache-Control", "public, max-age=3600");
        createReadStream(abs).pipe(res);
        return;
      } catch {
        res.status(404).json({ error: "File not found" });
        return;
      }
    }

    const file = await objectStorageService.searchPublicObject(filePath);
    if (!file) {
      res.status(404).json({ error: "File not found" });
      return;
    }

    const response = await objectStorageService.downloadObject(file);

    res.status(response.status);
    response.headers.forEach((value, key) => res.setHeader(key, value));

    if (response.body) {
      const nodeStream = Readable.fromWeb(response.body as ReadableStream<Uint8Array>);
      nodeStream.pipe(res);
    } else {
      res.end();
    }
  } catch (error) {
    req.log.error({ err: error }, "Error serving public object");
    res.status(500).json({ error: "Failed to serve public object" });
  }
});

/**
 * GET /storage/objects/*
 *
 * Serve object entities from PRIVATE_OBJECT_DIR.
 * These are served from a separate path from /public-objects and can optionally
 * be protected with authentication or ACL checks based on the use case.
 */
router.get("/storage/objects/*path", async (req: Request, res: Response) => {
  try {
    const raw = req.params.path;
    const wildcardPath = Array.isArray(raw) ? raw.join("/") : raw;
    const objectPath = `/objects/${wildcardPath}`;

    // Gate: only serve files actually referenced by a library entry
    // (either as the file storageRef or as a web-source screenshot).
    // This prevents arbitrary enumeration of the private bucket.
    const refMatch = `gcs:${objectPath}`;
    const referenced = await db
      .select({ id: libraryEntriesTable.id })
      .from(libraryEntriesTable)
      .where(
        or(
          eq(libraryEntriesTable.storageRef, refMatch),
          eq(libraryEntriesTable.storageRef, objectPath),
          eq(libraryEntriesTable.screenshotObjectPath, objectPath),
        ),
      )
      .limit(1);
    if (referenced.length === 0) {
      res.status(404).json({ error: "Not found" });
      return;
    }

    const objectFile = await objectStorageService.getObjectEntityFile(objectPath);

    const response = await objectStorageService.downloadObject(objectFile);

    res.status(response.status);
    response.headers.forEach((value, key) => res.setHeader(key, value));

    if (response.body) {
      const nodeStream = Readable.fromWeb(response.body as ReadableStream<Uint8Array>);
      nodeStream.pipe(res);
    } else {
      res.end();
    }
  } catch (error) {
    if (error instanceof ObjectNotFoundError) {
      req.log.warn({ err: error }, "Object not found");
      res.status(404).json({ error: "Object not found" });
      return;
    }
    req.log.error({ err: error }, "Error serving object");
    res.status(500).json({ error: "Failed to serve object" });
  }
});

export default router;